requirements.txt
google-cloud-storage==2.10.0
torchvision
opencv-python
xformers
runpod
